requirements.txt
gtts
elevenlabs
google-cloud-texttospeech
python-dotenv